backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
torch>=2.0.0
transformers>=4.30.0
albumentations>=1.3.0
Pillow>=10.0.0
numpy>=1.24.0
qdrant-client>=1.6.0
asyncpg>=0.28.0
python-dotenv>=1.0.0